tools/ed-search/pyproject.toml
[project]
name = "ed-search"
version = "0.1.0"
description = "Authenticated CLI for expireddomains.net — runs saved-search profiles, dumps CSV"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.1",
    "click>=8.1",
    "python-dotenv>=1.0",
    "platformdirs>=4.2",
]

[project.scripts]
ed-search = "ed_search.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ed_search"]
